modification et correction bugs

+++ tracé_de_rayon/raytracer/noeuds/Noeud.java
@@ -2,9 +2,9 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public class Noeud implements NoeudInterface {
-    
-    public Image calculer(SceneInterface scene, double x, double y, double l, double h) throws RemoteException{
-        return scene.compute(x,y,l,h);
+
+    public Image calculer(SceneInterface scene, int x, int y, int l, int h) throws RemoteException {
+        return scene.compute(x, y, l, h);
     }
-    
+
 }
